feat(geo-layers): Make Tileset2D less deck.gl dependent
