+++ modules/geo-layers/src/tileset-2d/index.ts
@@ -1,7 +1,6 @@
 // deck.gl, MIT license
 
 export type {
-  TileLoadProps,
   Bounds,
   ZRange,
   GeoBoundingBox,
@@ -9,16 +8,25 @@ export type {
   TileBoundingBox
 } from './types';
 
-export type {Tileset2DProps, RefinementStrategy} from './tileset-2d';
-export {Tileset2D, STRATEGY_DEFAULT} from './tileset-2d';
+export type {TileLoadProps} from './tile-2d-header';
+export type {TraversalParameters} from './tile-2d-traversal';
 
+export type {Tileset2DProps, RefinementStrategy, RefinementStrategyFunction} from './tileset-2d';
+export {Tileset2D} from './tileset-2d';
 export {Tile2DHeader} from './tile-2d-header';
 
 export {
-  isGeoBoundingBox,
-  isURLTemplate,
-  urlType,
-  getURLFromTemplate,
+  // isGeoBoundingBox,
+  tileIndexToBoundingBox,
+  /** @deprecated */
+  tileIndexToBoundingBox as tileToBoundingBox
+} from './tile-index-utils';
+
+export {
   getTileIndices,
-  tileToBoundingBox
-} from './utils';
+} from './get-tile-indices';
+
+export {
+  isURLTemplate,
+  getURLFromTemplate
+} from './url-utils';
