+++ test/modules/geo-layers/tile-3d-layer.spec.js
