--- test/modules/geo-layers/tile-3d-layer/index.js
@@ -1,1 +1,0 @@
-import './tile-3d-layer.spec';
