+++ test/modules/geo-layers/tile-layer.spec.js
