--- test/modules/geo-layers/tile-layer/index.js
@@ -1,1 +1,0 @@
-import './tile-layer.spec';
