--- test/modules/geo-layers/tileset-2d/index.js
@@ -1,3 +1,0 @@
-import './utils.spec';
-import './tileset-2d.spec';
-import './tile-2d-header.spec';
