+++ test/modules/geo-layers/tileset-2d/index.ts
@@ -1,0 +1,6 @@
+// deck.gl, MIT license
+
+import './url-utils.spec';
+import './tileset-2d-traversal.spec';
+import './tileset-2d.spec';
+import './tile-2d-header.spec';
