+++ test/modules/geo-layers/tileset-2d/tileset-2d-traversal.spec.ts
